+++ app/models/claim.rb
@@ -88,6 +88,8 @@ class Claim < ApplicationRecord
     male: 2,
   }
 
+  validates :academic_year, format: {with: PolicyConfiguration::ACADEMIC_YEAR_REGEXP}
+
   validates :payroll_gender, on: [:gender, :submit], presence: {message: "Choose the option for the gender your school’s payroll system associates with you"}
 
   validates :first_name, on: [:name, :submit], presence: {message: "Enter your first name"}